+++ src/commands/tables/index.ts
@@ -3,3 +3,4 @@ export * from './scripts';
 export * from './maintenance';
 export * from './profile';
 export * from './definition';
+export * from './export';